requirements.txt
opencv-python
pyzbar
pyautogui
pygetwindow
pyyaml
simpleaudio
PyQt6